+++ react/src/components/Table/index.js
@@ -1,0 +1,1 @@
+export { Table, TableWrap, TableCell, TableCellAction, KebabIcon, default } from './Table';